Find points to remove accurately
Batching both the addition and removal processes seemed natural, and was
a speed-up over a one-by-one iteration for both, but is not necessary
for the removal process.

The addition process benefits from batching because it's pool is so big
and expensive to calculate (all possible line segments between all
points of all possible colors: 125,000 combinations for 250 points & 4
colors). Batching allows us to calculate that pool (and their scores)
once for many strings. The scores are not particularly accurate,
however, since they do not take into account the other strings also
added, but gradually reducing the batching size helps. Similar to
simulated annealing, where the reducing batch size is similar to the
reducing simulated temperature.

On the other hand, the removal process does not benefit as much from
batching. The pool of strings to remove is never very large (most images
stay below 3000 strings, more detailed images can have up to 4000). It's
more beneficial to remove strings accurately.

Without this change, addition & removal batches tended to cancel each
other out a bunch. It was possible to see the same strings added and
removed over and over. Now, because the removals are accurate, this
back-and-forth action doesn't happen.

This change reduced the time for my test from about 120s to about 50s on
a 605x605px colored image. The relevant runtime params were:

    --step-size 0.7 --string-alpha 0.4 --pin-count 100 --pin-arrangement random --style auto-color