Revert 93c11a3, but keep the spirit of the change
The point of that change was that removing points accurately is easier
than adding points accurately, so they don't need to be batched as
aggressively. However, that change removed batching entirely, which can
still be beneficial. This change restores batched removal, but does it
much less aggressively.

This test reduced the time for my test from about 38min to about 20min
on a 640x640px b&w image. The relevant runtime params were:

    --step-size 1 --string-alpha 0.05 --pin-count 600 --pin-arrangement circle --style black-on-white