Use a nicer RefImage/scoring model
The RefImage has two responsibilities:

- Keep track of the selected strings & score that collection
- Produce an image from a collection of strings

This change simplifies the first responsibility.

Before, the scoring model used a white image as the "perfect" score.
Now, the "perfect" image is black. This makes some of the math simpler.
Also, it means that there is a definitive simple/default case: when
drawing with white strings on a black background neither the image nor
the strings have to be inverted.

When scoring, the RefImage is thought of as a "hole" that needs filling.
The RefImage starts out as a negated copy of the original
image---negated meaning each pixels RGB values have been made negative,
ie a white pixel `RGB { 255, 255, 255 }` has been transformed into `RGB
{ -255, -255, -255 }`. The program then adds in white strings to "fill
in" the negative hole and try to approximate a level ground of `0` for
each pixel's RGB values.

For black-on-white images, the image & string colors need to be
inverted, and then the process is identical.

(TODO: separate out the two responsibilities of the RefImage into two
different structs.)